+++ docs/CHANGES.rst
@@ -25,6 +25,9 @@ Changelog
   ``api.user.revoke_roles``
   [adamcheasley]
 
+- Added ``api.folder.list_objects()`` and ``api.folder.list_brains()``
+  [ajung]
+
 1.2.1 (2014-06-24)
 ------------------